requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dataclasses-json==0.6.4
deepgram-sdk==3.0.1
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.0
Flask==2.0.3
Flask-Cors==4.0.0
frozenlist==1.4.1
gotrue==2.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jwt==1.3.1
MarkupSafe==2.1.5
marshmallow==3.21.0
multidict==6.0.5
mypy-extensions==1.0.0
openai==1.13.3
packaging==23.2
postgrest==0.16.1
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
storage3==0.7.0
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
verboselogs==1.7
websockets==11.0.3
Werkzeug==2.2.2
yarl==1.9.4
gunicorn==20.0.4
slack_sdk
PyPDF2
google-api-python-client

